+++ packages/react-cards/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui/react-cards
 
-This log was last generated on Wed, 13 Nov 2024 07:08:33 GMT and should not be manually modified.
+This log was last generated on Thu, 14 Nov 2024 01:04:05 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [0.205.186](https://github.com/microsoft/fluentui/tree/@fluentui/react-cards_v0.205.186)
+
+Thu, 14 Nov 2024 01:04:05 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-cards_v0.205.185..@fluentui/react-cards_v0.205.186)
+
+### Patches
+
+- Bump @fluentui/react to v8.121.11 ([PR #33268](https://github.com/microsoft/fluentui/pull/33268) by beachball)
 
 ## [0.205.185](https://github.com/microsoft/fluentui/tree/@fluentui/react-cards_v0.205.185)