+++ packages/react-docsite-components/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui/react-docsite-components
 
-This log was last generated on Wed, 13 Nov 2024 07:08:33 GMT and should not be manually modified.
+This log was last generated on Thu, 14 Nov 2024 01:04:05 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [8.13.135](https://github.com/microsoft/fluentui/tree/@fluentui/react-docsite-components_v8.13.135)
+
+Thu, 14 Nov 2024 01:04:05 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-docsite-components_v8.13.134..@fluentui/react-docsite-components_v8.13.135)
+
+### Patches
+
+- Bump @fluentui/react to v8.121.11 ([PR #33268](https://github.com/microsoft/fluentui/pull/33268) by beachball)
+- Bump @fluentui/react-monaco-editor to v1.7.253 ([PR #33268](https://github.com/microsoft/fluentui/pull/33268) by beachball)
 
 ## [8.13.134](https://github.com/microsoft/fluentui/tree/@fluentui/react-docsite-components_v8.13.134)